% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anDiagnostic_report.R
\name{anDiagnostic_report}
\alias{anDiagnostic_report}
\title{ANALYSIS - Compare and diagnose estimates.}
\usage{
anDiagnostic_report(
  estimates,
  attribute,
  small_area = "DOMAIN",
  large_area,
  outfolder = NULL
)
}
\arguments{
\item{estimates}{List containing FIESTA estimation output.}

\item{attribute}{String. Name of estimated attribute, should be the name 
used in `data` containing the estimates you would like to compare.}

\item{small_area}{String. Column name in `multest` containing small area of
interest codes. Defaults to "DOMAIN"}

\item{large_area}{String. Column name in `multest` containing large area that 
strength was borrowed out to.}

\item{outfolder}{String. The path of folder to output tables.}
}
\description{
Generates a report for estimator comparisions and diagnoses.
}
\details{
If variables are NULL, then it will prompt user to input variables.
}
\author{
Tracey S. Frescino, Grayson W. White
}
\keyword{data}
