% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGBest_core.R
\name{anGBest_core}
\alias{anGBest_core}
\title{Generate core tables using Green-Book module.}
\usage{
anGBest_core(
  state = NULL,
  datsource = "datamart",
  data_dsn = NULL,
  evalCur = TRUE,
  evalEndyr = NULL,
  fill = TRUE,
  addSEcol = FALSE,
  allin1 = TRUE,
  savedata = FALSE,
  outfolder = NULL,
  outfn.date = TRUE,
  estround = 2,
  pseround = 2,
  divideby.vol = "million",
  datCore = NULL
)
}
\arguments{
\item{state}{String. Name of a state for tables (e.g., 'Utah')}

\item{datsource}{String. Source of data ('datamart', 'sqlite').}

\item{data_dsn}{String. If datsource='sqlite', the name of SQLite database
(*.sqlite).}

\item{evalCur}{Logical. If TRUE, uses the most current FIA Evaluation in
FIA DataMart for the state.}

\item{evalEndyr}{Integer. Last year of the FIA Evaluation of interest.}

\item{fill}{Logical. If TRUE, color fill is applied to all tables.}

\item{addSEcol}{Logical. If TRUE, percent standard error values are added to
columns side-by-side.}

\item{allin1}{Logical. If TRUE, percent standard error values are added to
same cell as estimates.}

\item{savedata}{Logical. If TRUE, saves data objects to outfolder (*.rda)}

\item{outfolder}{String. The path of folder to output tables.}

\item{outfn.date}{Logical. If TRUE, appends current date to outfile name.}

\item{estround}{Integer. Number of digits after decimal to round estimate to.}

\item{pseround}{Integer. Number of digits after decimal to round percent standard error to.}

\item{divideby.vol}{String. divideby for volume tables ('hundred',
'thousand', 'million').}

\item{datCore}{R list object. FIA data from database, returned from
FIESTAnalysis::anGBest_core_data. Includes datPlots, datStrata, ref_countycd, 
and ref_unitcd.}
}
\description{
Generates a set of core tables for a state.
}
\details{
If variables are NULL, then it will prompt user to input variables.
}
\examples{

# Run a set of core tables for Utah, for the most current FIA Evaluation. 
# Note: this may take a few minutes

# anGBest_core("Utah", evalCur=TRUE)


}
\author{
Tracey S. Frescino
}
\keyword{data}
