% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGBest_core_data.R
\name{anGBest_core_data}
\alias{anGBest_core_data}
\title{Extract data for core table from FIA DataMart.}
\usage{
anGBest_core_data(
  state,
  datsource = "datamart",
  data_dsn = NULL,
  evalCur = TRUE,
  evalEndyr = NULL,
  evalType = c("CURR", "VOL", "GRM"),
  savedata = TRUE,
  outfolder = NULL,
  outfn.date = TRUE
)
}
\arguments{
\item{state}{String. Name of a state for tables (e.g., 'Utah')}

\item{datsource}{String. Source of data ('datamart', 'sqlite').}

\item{data_dsn}{String. If datsource='sqlite', the name of SQLite database
(*.sqlite).}

\item{evalCur}{Logical. If TRUE, uses the most current FIA Evaluation in
FIA DataMart for the state.}

\item{evalEndyr}{Integer. Last year of the FIA Evaluation of interest.}

\item{evalType}{String vector. The type(s) of evaluation of interest ('ALL',
'CURR', 'VOL', 'GRM', 'P2VEG', 'DWM", 'INV', 'REGEN', 'CRWN').  The evalType
'ALL' includes nonsampled plots; 'CURR' includes plots used for area
estimates; 'VOL' includes plots used for area and/or tree estimates; The
evalType 'GRM' includes plots used for growth, removals, mortality, and
change estimates (eval_typ %in% c(GROW, MORT, REMV, CHNG)).  Multiple types
are accepted. See details below and FIA database manual for regional
availability and/or differences. Note: do not use if EVALID is specified.}

\item{savedata}{Logical. If TRUE, saves data objects to outfolder (*.rda)}

\item{outfolder}{String. The path of folder to output tables.}

\item{outfn.date}{Logical. If TRUE, appends current date to outfile name.}
}
\description{
Extract data from FIA Datamart for input to FIESTAnalysis::anGBest_core
function. The function generates a set of core tables for a state 
(https://apps.fs.usda.gov/fia/datamart/datamart.html).
}
\details{
If variables are NULL, then it will prompt user to input variables.
}
\author{
Tracey S. Frescino
}
\keyword{data}
