% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGBgetpop.R
\name{anGBgetpop}
\alias{anGBgetpop}
\title{Analysis - Generate estimates for custom-defined boundary.}
\usage{
anGBgetpop(
  GBdata = NULL,
  popType = "VOL",
  showsteps = FALSE,
  returnxy = TRUE,
  savedata = FALSE,
  saveobj = FALSE,
  savedata_opts = NULL,
  ...
)
}
\arguments{
\item{GBdata}{R list object. FIA data for estimation.}

\item{popType}{String. Type of evaluation(s) to include in population data
("VOL"). Note: currently only 'VOL' is available.}

\item{showsteps}{Logical. If TRUE, show intermediate steps.}

\item{returnxy}{Logical. If TRUE, saves xy coordinates to outfolder.}

\item{savedata}{Logical. If TRUE, saves data objects to outfolder.}

\item{saveobj}{Logical. If TRUE, saves GBpopdat object to outfolder.}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options. Only used when savedata = TRUE.}

\item{...}{Parameters for GBdata.}
}
\value{
Estimate.
}
\description{
Wrapper to get plot data, stratification data and estimates within a
custom-defined boundary. Includes FIESTA functions: anGBdata()
(spGetPlots(), spGetStrata()), and modGBpop().
}
\author{
Tracey S. Frescino
}
\keyword{data}
