% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGetAuxiliary_list.R
\name{anGetAuxiliary_list}
\alias{anGetAuxiliary_list}
\title{Data - Get list of XY data by bnd.}
\usage{
anGetAuxiliary_list(
  xydatlst,
  saveobj = FALSE,
  objnm = "auxdatlst",
  outfolder = NULL,
  overwrite = FALSE,
  outfn.pre = NULL,
  outfn.date = FALSE,
  ...
)
}
\arguments{
\item{xydatlst}{sf R object. List output from FIESTAnalysis::anGetXY_list function.}

\item{saveobj}{Logical. If TRUE, saves xydatlst object to outfolder.}

\item{objnm}{String. If savedata=TRUE, name of object to save.}

\item{outfolder}{String. Name of outfolder. If NULL, saves to working directory.}

\item{overwrite}{Logical. If TRUE, overwrite object in outfolder.}

\item{outfn.pre}{String. If saveobj=TRUE, prefix for object name.}

\item{outfn.date}{Logical. If TRUE, add current date to object name.}

\item{...}{Parameter for FIESTA::spGetAuxiliary function.}
}
\value{
Data.
}
\description{
Wrapper to get XY data.
}
\author{
Tracey S. Frescino
}
\keyword{data}
