% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGetData_list.R, R/anGetPlots_list.R
\name{anGetData_list}
\alias{anGetData_list}
\title{Data - Get list of plot data and auxiliary data for estimation.}
\usage{
anGetData_list(
  bndlst,
  splitbnd = TRUE,
  splitby = 100,
  savelistobj = TRUE,
  listoutfolder = NULL,
  listobjnm = "datalst",
  ...
)

anGetData_list(
  bndlst,
  splitbnd = TRUE,
  splitby = 100,
  savelistobj = TRUE,
  listoutfolder = NULL,
  listobjnm = "datalst",
  ...
)
}
\arguments{
\item{bndlst}{sf R object. List object of boundary extent(s).}

\item{splitbnd}{Logical. If TRUE, splits list into multiple lists.}

\item{splitby}{numeric. If splitbnd=TRUE, the number of domains to split by.}

\item{savelistobj}{Logical. If TRUE, saves listobjects to listoutfolder.}

\item{listoutfolder}{NULL. Folder to save list object. If NULL, saves 
to working directory.}

\item{listobjnm}{String. Name of saved object.}

\item{...}{Other parameters for anGetData().}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options. Only used when savedata = TRUE or saveobj = TRUE.}
}
\value{
Data.

Data.
}
\description{
Wrapper to get plot data and auxiliary data for estimation.

Wrapper to get plot data and auxiliary data for estimation.
}
\author{
Tracey S. Frescino
}
\keyword{data}
