% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGetDEM_derivatives.R
\name{anGetDEM_derivatives}
\alias{anGetDEM_derivatives}
\title{Spatial analysis - Creates derivative rasters from a digital elevation model (DEM).}
\usage{
anGetDEM_derivatives(
  demfn,
  derivatives = c("slope", "aspect", "roughness", "TRI", "TPI"),
  aspect.derivatives = TRUE,
  aspect.keep = FALSE,
  compress = "DEFLATE",
  BigTIFF = FALSE,
  outfolder = NULL,
  outfn.pre = "dem",
  outext = "tif",
  overwrite = FALSE
)
}
\arguments{
\item{demfn}{String or Raster. File name of DEM.}

\item{derivatives}{String vector. List of derivatives to calculate from
DEM ('slope', 'aspect', 'roughness', 'TRI', 'TPI').}

\item{aspect.derivatives}{Logical. If TRUE, calculate northing and 
easting derivatives.}

\item{aspect.keep}{Logical. If aspect.derivatives = TRUE, if TRUE, 
returns aspect in list.}

\item{compress}{String. An optional compression type ('LZW', "DEFLATE',
"PACKBITS').}

\item{BigTIFF}{Logical. If TRUE, adds BIGTIFF=YES to export (for files > 4GB).}

\item{outfolder}{String. If exportsp=TRUE, name of output folder. If NULL,
the working directory is used.}

\item{outfn.pre}{String. Name of output prefix for raster derivative(s). 
If NULL, default is 'dem'.}

\item{outext}{String. Name of raster extension (fmt). If NULL, uses
extension from outfn or rastfn.}

\item{overwrite}{Logical. If TRUE, overwrites raster derivatives.}
}
\value{
\item{rasterlst}{ String list. Full path name of raster files. }
}
\description{
Creates one or more derivative rasters from digital elevation model (DEM)
using gdalUtils::gdaldem(). The output will have the same extent, cell size, 
and coordinate reference system as the input DEM. See below for list of 
derivatives to choose from.
}
\details{
DEM Derivatives:\cr 
\tabular{ll}{ 
\tab slope, in degrees\cr 
\tab aspect, in degrees\cr 
\tab roughness\cr
\tab TRI (Terrain Ruggedness Index)\cr
\tab TPI (Terrain Position Index)\cr
\tab northness - aspect northness\cr
\tab eastness - aspect eastness\cr }
}
\author{
Tracey S. Frescino, Chris Toney
}
\keyword{spatial}
