% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGetPixelData.R
\name{anGetPixelData}
\alias{anGetPixelData}
\title{Data - Extract pixel-level data.}
\usage{
anGetPixelData(
  ref.rastfn,
  bnd,
  bnd_dsn = NULL,
  bnd.att = NULL,
  bnd.filter = NULL,
  bnd.byvar = NULL,
  polyvlst = NULL,
  polyvarnmlst = NULL,
  rastfolder = NULL,
  rastlst = NULL,
  rastlst.name = NULL,
  returnxy = FALSE,
  exportsp = FALSE,
  savedata_opts = NULL,
  ncores = 1
)
}
\arguments{
\item{ref.rastfn}{String. Path name of raster to use as reference for
pixel size.}

\item{bnd}{sf R object or String. Boundary layer of estimation domain
units. Can be a spatial sf object, full pathname to a shapefile, or name of
a layer within a database.}

\item{bnd_dsn}{String. Name of data source name with bnd_layer, if in a
database.}

\item{bnd.att}{String. Name of attribute in boundary file to define
estimation domain units.}

\item{bnd.filter}{String. Optional filter of bnd_layer.}

\item{bnd.byvar}{String. Optional name in bnd to get pixel data by.}

\item{polyvlst}{String vector or list. The name(s) of variable(s) to
extract from polygon(s). If extracting multiple variables from more than one
polygon, specify names in a list format, corresponding to polyvlst.}

\item{polyvarnmlst}{String vector or list. Output name(s) of variable(s)
extracted from polygon(s). If extracting multiple variables from more than
one polygon, specify names in a list format, corresponding to polyvlst. The
number of names must match the number of variables in polyvarlst.}

\item{rastfolder}{String. Name of the folder with raster layers. Optional.
Useful if all raster layers are in same folder.}

\item{rastlst}{String vector or list. A list of raster(s) with
continuous data and/or categorical values (e.g., DEM). The list may include 
file name of raster(s) or raster objects that are not InMemory.}

\item{rastlst.name}{String vector. Output names for rastlst.
Optional. If NULL, name of raster is used as default or name+'_'+layer
number for multi-band layers.}

\item{returnxy}{Logical. If TRUE, returns XY coordinates.
includes XY data if returnxy = TRUE.}

\item{exportsp}{Logical. If TRUE, the extracted raster point data are
exported to outfolder.}

\item{savedata_opts}{List. See help(savedata_options()) for a list
of options. Only used when savedata = TRUE.}

\item{ncores}{R object. Number of cores to use for data extraction.
(Note: needs parallel package).}
}
\value{
Data.
}
\description{
Wrapper to get pixel data, including auxiliary information.
}
\note{
If exportsp=TRUE:\cr If out_fmt="shp", the writeOGR (rgdal) function is
called. The ArcGIS driver truncates variable names to 10 characters or less.
Variable names are changed before export using an internal function.
If Spatial o bject has more than 1 record, it will be returned but not exported.

The spTransform (rgdal) method is used for on-the-fly map projection
conversion and datum transformation using PROJ.4 arguments. Datum
transformation only occurs if the +datum tag is present in the both the from
and to PROJ.4 strings. The +towgs84 tag is used when no datum transformation
is needed. PROJ.4 transformations assume NAD83 and WGS84 are identical
unless other transformation parameters are specified.  Be aware, providing
inaccurate or incomplete CRS information may lead to erroneous data shifts
when reprojecting. See spTransform help documentation for more details.
}
\author{
Tracey S. Frescino
}
\keyword{data}
