% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGetPlots_list.R
\name{anGetPlots_list}
\alias{anGetPlots_list}
\title{Data - Get list of XY data by bnd.}
\usage{
anGetPlots_list(
  auxdatlst,
  saveobj = FALSE,
  objnm = "pltdatlst",
  outfolder = NULL,
  overwrite = FALSE,
  outfn.pre = NULL,
  outfn.date = FALSE,
  ...
)
}
\arguments{
\item{auxdatlst}{sf R object. List output from FIESTAnalysis::anGetAuxiliary_list function.}

\item{saveobj}{Logical. If TRUE, saves xydatlst object to outfolder.}

\item{objnm}{String. If savedata=TRUE, name of object to save.}

\item{outfolder}{String. Name of outfolder. If NULL, saves to working directory.}

\item{overwrite}{Logical. If TRUE, overwrite object in outfolder.}

\item{outfn.pre}{String. If saveobj=TRUE, prefix for object name.}

\item{outfn.date}{Logical. If TRUE, add current date to object name.}

\item{...}{Parameter for FIESTA::spGetPlots function.}
}
\value{
Data.
}
\description{
Wrapper to get XY data.
}
\author{
Tracey S. Frescino
}
\keyword{data}
