% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anGetXY_list.R
\name{anGetXY_list}
\alias{anGetXY_list}
\title{Data - Get list of XY data by bnd.}
\usage{
anGetXY_list(
  bndlst,
  saveobj = FALSE,
  objnm = "xydatlst",
  outfolder = NULL,
  overwrite = FALSE,
  outfn.pre = NULL,
  outfn.date = FALSE,
  ...
)
}
\arguments{
\item{bndlst}{sf R object. List of bnd sf objects.}

\item{saveobj}{Logical. If TRUE, saves xydatlst object to outfolder.}

\item{objnm}{String. If savedata=TRUE, name of object to save.}

\item{outfolder}{String. Name of outfolder. If NULL, saves to working directory.}

\item{overwrite}{Logical. If TRUE, overwrite object in outfolder.}

\item{outfn.pre}{String. If saveobj=TRUE, prefix for object name.}

\item{outfn.date}{Logical. If TRUE, add current date to object name.}

\item{...}{Parameter for FIESTA::spGetXY function.}
}
\value{
Data.
}
\description{
Wrapper to get XY data.
}
\author{
Tracey S. Frescino
}
\keyword{data}
