% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anMApop_list.R
\name{anMApop_list}
\alias{anMApop_list}
\title{Data - Get list of population data for estimation.}
\usage{
anMApop_list(
  datalst,
  prednames = NULL,
  saveobj = FALSE,
  objnm = "MApopdat",
  outfolder = NULL,
  overwrite = FALSE,
  outfn.pre = NULL,
  outfn.date = FALSE,
  ...
)
}
\arguments{
\item{datalst}{sf R object. List object output from FIESTAnalysis::anGetData_list().}

\item{prednames}{String vector. One or more predictors to use for estimation.}

\item{saveobj}{Logical. If TRUE, saves SApopdat object to outfolder.}

\item{objnm}{String. If savedata=TRUE, name of object to save.}

\item{outfolder}{String. Name of outfolder. If NULL, saves to working directory.}

\item{overwrite}{Logical. If TRUE, overwrite object in outfolder.}

\item{outfn.pre}{String. If saveobj=TRUE, prefix for object name.}

\item{outfn.date}{Logical. If TRUE, add current date to object name.}

\item{...}{Additional parameters to FIESTA::modMApop}
}
\value{
Data.
}
\description{
Wrapper to get population data for estimation.
}
\author{
Tracey S. Frescino
}
\keyword{data}
