% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anMApop_report.R
\name{anMApop_report}
\alias{anMApop_report}
\title{ANALYSIS - Generate a report for Green-Book custom population.}
\usage{
anMApop_report(
  MApopdat,
  MAmethod = "greg",
  AOInm,
  photofn = NULL,
  photo_author = NULL,
  pcfilter = NULL,
  spcd = NULL,
  title.ref = NULL,
  totals = TRUE,
  outfolder
)
}
\arguments{
\item{MApopdat}{List object. Population data from modMApop().}

\item{MAmethod}{String. mase (i.e., model-assisted) method to use 
('greg', 'gregEN', 'ratio').}

\item{AOInm}{String. Name of area of interest.}

\item{photofn}{String. Path name to photo to add to report.}

\item{photo_author}{String. Name of photo author (Optional).}

\item{pcfilter}{String. Name of variable or string to filter plots or
conditions.  Must be in R syntax (e.g., ADFORCD == 408, COUNTYCD == 1).
Optional.}

\item{spcd}{Integer. A species code to use as focus species of the report.
Multiple codes are accepted if the focus species is a combined group.
Default is 746 (quaking aspen).}

\item{title.ref}{String. Reference information for title and output file
names.}

\item{totals}{Logical. If TRUE, per acre estimates are multiplied by acres.}

\item{outfolder}{String. The path of folder to output tables.}
}
\description{
Generates a standard report based on area of interest.
}
\details{
If variables are NULL, then it will prompt user to input variables.
}
\author{
Tracey S. Frescino
}
\keyword{data}
