% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anPBpopICE_core.R
\name{anPBpopICE_core}
\alias{anPBpopICE_core}
\title{ANALYSIS - Generate core tables for ICE.}
\usage{
anPBpopICE_core(
  PBpopdatICE,
  AOInm,
  T1,
  T2,
  tabtype = "PCT",
  sumunits = FALSE,
  pnt1 = FALSE,
  domlut = NULL,
  outfolder = NULL,
  outfn.pre = NULL,
  outfn.date = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{PBpopdatICE}{List object. Population data output from
PBpopdatICE().}

\item{AOInm}{String. The name of the area of interest (i.e., population) for
titles and output names.}

\item{T1}{Integer. Year of Time 1 imagery (YYYY).}

\item{T2}{Integer. Year of Time 2 imagery (YYYY).}

\item{tabtype}{String. Type of units for the table ("AREA", "PCT").}

\item{sumunits}{Logical. If TRUE, estimation units are summed and returned
in one table.}

\item{pnt1}{Logical. If TRUE, use only center point when no change observed.}

\item{domlut}{Data frame. Lookup table of domain categories and titles for
estimation.  If NULL, default from PBpopdatICE().}

\item{outfolder}{String. The path of folder to output tables.}

\item{outfn.pre}{String. Add a prefix to output name (e.g., "01").}

\item{outfn.date}{Logical. If TRUE, adds current date to outfile name.}

\item{overwrite}{Logical. If TRUE, overwrites file if it exists.}
}
\value{
The following tables are saved to a folder named 'estimates' in the
outfolder.  If rawdata=TRUE, the rawdata is saved to a folder named
'rawdata' in the estimates folder.

\item{table01}{ Estimates of change_pnt on all lands. } \item{table02}{
Estimates of chg_ag_2_GRP on all lands. } \item{table03}{ Estimates of land
use at Time 1 on all lands. } \item{table04}{ Estimates of land use at Time
2 on all lands. } \item{table05}{ Estimates of land cover at Time 1 all
lands. } \item{table06}{ Estimates of land cover at Time 2 on all lands. }
\item{table07}{ Percent of forest land use change. } \item{table08}{ Percent
of land cover change where land use was forest at Time 1. } \item{table09}{
Percent of land use change from Time 1 to Time 2. } \item{table10}{ Percent
of land cover change from Time 1 to Time 2. } \item{table11}{ Percent of
land cover by land use at Time 1. } \item{table12}{ Percent of land cover by
land use at Time 2. } \item{table13}{ Percent change of land use at Time 1
within land use at Time 2. } \item{table13a}{ Percent change of land use at
Time 1 within land cover at Time 2. } \item{table14}{ Percent change of land
cover at Time 1 within land use at Time 2. } \item{table15}{ Percent change
of land use at Time 1 within within change agent group. } \item{table16}{
Percent change of land cover at Time 1 within change agent group. }
\item{table17}{ Percent change of Land Use from Time 1 to Time 2. }
\item{table18}{ Percent change of Land Cover from Time 1 to Time 2. }
\item{table19}{ Percent change of Land Use (Forest/Nonforest) from Time 1 to
Time 2. }
}
\description{
Generates a set of core tables based on area of interest.
}
\details{
If variables are NULL, then it will prompt user to input variables.
}
\references{
Frescino, Tracey S.; Moisen, Gretchen G.; Megown, Kevin A.;
Nelson, Val J.; Freeman, Elizabeth A.; Patterson, Paul L.; Finco, Mark;
Brewer, Ken; Menlove, James 2009.  Nevada Photo-Based Inventory Pilot (NPIP)
photo sampling procedures. Gen. Tech. Rep.  RMRS-GTR-222. Fort Collins, CO:
U.S. Department of Agriculture, Forest Service, Rocky Mountain Research
Station. 30 p.

Patterson, Paul L. 2012. Photo-based estimators for the Nevada photo-based
inventory.  Res. Pap. RMRS-RP-92. Fort Collins, CO: U.S. Department of
Agriculture, Forest Service, Rocky Mountain Research Station. 14 p.
}
\author{
Tracey S. Frescino
}
\keyword{data}
