% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anSApop_list.R
\name{anSApop_list}
\alias{anSApop_list}
\title{Data - Get list of plot data and auxiliary data for estimation.}
\usage{
anSApop_list(
  SAdatalst,
  smallbndlst,
  smallbnd.domain,
  prednames = NULL,
  saveobj = FALSE,
  objnm = "SApopdat",
  outfolder = NULL,
  overwrite = FALSE,
  outfn.pre = NULL,
  outfn.date = FALSE
)
}
\arguments{
\item{SAdatalst}{sf R object. List object output from FIESTAnalysis::anGetData_list().}

\item{smallbndlst}{sf R object. List object of smallbnds associated with SAdatalst.}

\item{smallbnd.domain}{String. Name of unique identifier for smallbnd.}

\item{prednames}{String vector. One or more predictors to use for estimation.}

\item{saveobj}{Logical. If TRUE, saves SApopdat object to outfolder.}

\item{objnm}{String. If savedata=TRUE, ame of object to save.}

\item{outfolder}{String. Name of outfolder. If NULL, saves to working directory.}

\item{overwrite}{Logical. If TRUE, overwrite object in outfolder.}

\item{outfn.pre}{String. If saveobj=TRUE, prefix for object name.}

\item{outfn.date}{Logical. If TRUE, add current date to object name.}
}
\value{
Data.
}
\description{
Wrapper to get plot data and auxiliary data for estimation.
}
\note{
If exportsp=TRUE:\cr If out_fmt="shp", the writeOGR (rgdal) function is
called. The ArcGIS driver truncates variable names to 10 characters or less.
Variable names are changed before export using an internal function
If Spatial o bject has more than 1 record, it will be returned but not exported.

The spTransform (rgdal) method is used for on-the-fly map projection
conversion and datum transformation using PROJ.4 arguments. Datum
transformation only occurs if the +datum tag is present in the both the from
and to PROJ.4 strings. The +towgs84 tag is used when no datum transformation
is needed. PROJ.4 transformations assume NAD83 and WGS84 are identical
unless other transformation parameters are specified.  Be aware, providing
inaccurate or incomplete CRS information may lead to erroneous data shifts
when reprojecting. See spTransform help documentation for more details.
}
\author{
Tracey S. Frescino
}
\keyword{data}
