% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anSApop_report.R
\name{anSApop_report}
\alias{anSApop_report}
\title{ANALYSIS - Generate a report small area population.}
\usage{
anSApop_report(
  SApopdatlst,
  SApackage,
  SAmethod,
  AOInm,
  pcfilter = NULL,
  fortypgrpcd = NULL,
  title.ref = NULL,
  totals = TRUE,
  outfolder = NULL
)
}
\arguments{
\item{SApopdatlst}{List object. Population data from modSApop().}

\item{SApackage}{String. Small area package to use ('JoSAE', 'sae')}

\item{SAmethod}{String. Small area method to use ('unit', 'area')}

\item{AOInm}{String. Name of area of interest.}

\item{pcfilter}{String. Name of variable or string to filter plots or
conditions.  Must be in R syntax (e.g., ADFORCD == 408, COUNTYCD == 1).
Optional.}

\item{fortypgrpcd}{Integer. A forest type group code to use as focus in the
report.}

\item{title.ref}{String. Reference information for title and output file
names.}

\item{totals}{Logical. If TRUE, per acre estimates are multiplied by acres.}

\item{outfolder}{String. The path of folder to output tables.}
}
\description{
Generates a standard report based on area of interest.
}
\details{
If variables are NULL, then it will prompt user to input variables.
}
\author{
Tracey S. Frescino
}
\keyword{data}
