% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEVALIDator_compare.R
\name{checkpop}
\alias{checkpop}
\title{Check FIADB and FIESTA population data}
\usage{
checkpop(FIADBpop, FIESTApop, evalendType = "01", rnd = 10)
}
\arguments{
\item{FIADBpop}{population data FIADB}

\item{FIESTApop}{population data FIESTA}

\item{evalendType}{String. last two characters of the evalid identifying the type of response}

\item{rnd}{Integer. rnd}
}
\description{
Check FIADB and FIESTA population data
}
