% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEVALIDator_compare.R
\name{compareAPI}
\alias{compareAPI}
\title{Compare outputs from EVALIDator and FIESTA}
\usage{
compareAPI(
  EVALIDatorlst,
  FIESTAlst,
  compareType = "ROW",
  precision_threshold = 1
)
}
\arguments{
\item{EVALIDatorlst}{List of estimates output from EVALIDator}

\item{FIESTAlst}{List of estimates output from FIESTA}

\item{compareType}{String ('TOTALS', 'ROW', 'COL', 'GRP')}

\item{precision_threshold}{Numeric. How close do estimates have to be to be considered a match?}
}
\description{
Compare outputs from EVALIDator and FIESTA
}
