% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEVALIDator_compare.R
\name{getAPIest}
\alias{getAPIest}
\title{Get estimate from EVALIDator API}
\usage{
getAPIest(
  evalid,
  estvar = NULL,
  estvar.filter = NULL,
  landarea = "FOREST",
  rowvar = NULL,
  colvar = NULL,
  ratio = FALSE,
  ratiotype = "PERACRE",
  estvard = NULL,
  estvard.filter = NULL,
  chng_type = "total",
  chng_measurements = "both",
  woodland = "Y",
  saplings = FALSE,
  dia5inch = FALSE,
  dsn = NULL,
  conn = NULL,
  attribute_nbr = NULL,
  attribute_nbrd = NULL,
  strFilter = NULL,
  EVALIDATOR_POP_ESTIMATE = NULL
)
}
\arguments{
\item{evalid}{6 digit Evaluation ID}

\item{estvar}{estimation variable}

\item{estvar.filter}{estimation filter ('TREECLCD = 2', 'STATUSCD = 2', 'STATUSCD = 1'). Similar to EVALIDator estn_type ('GS'-growing-stock; 'STGD'-standing-dead; 'AL'-all live)}

\item{landarea}{('FOREST', 'TIMBERLAND', 'ALL'). Similar to EVALIDator land_basis ('Forest land', Timberland', 'All land')}

\item{rowvar}{row domain to estimate by}

\item{colvar}{column domain to estimate by}

\item{ratio}{if ratio estimation}

\item{ratiotype}{type of ratio ('PERACRE', 'PERTREE')}

\item{estvard}{estimation variable for denominator}

\item{estvard.filter}{estimation filter for denominator ('TREECLCD = 2', 'STATUSCD = 2', 'STATUSCD = 1'). Similar to EVALIDator estd_type ('GS'-growing-stock; 'STGD'-standing-dead; 'AL'-all live)}

\item{chng_type}{c('total', 'annual')}

\item{chng_measurements}{('both', 'either')}

\item{woodland}{('Y', 'N', 'only') - whether to include woodland tree species}

\item{saplings}{Saplings.}

\item{dia5inch}{Logical. If TRUE, gets attribute with 5 inches and above}

\item{dsn}{data source name of SQLite database with EVALIDATOR_POP_ESTIMATE table (if attribute_nbr is NULL)}

\item{conn}{open connection of SQLite database with EVALIDATOR_POP_ESTIMATE table (if attribute_nbr is NULL)}

\item{attribute_nbr}{EVALIDator attribute number for estimation variable. Using attribute_nbr replaces: estvar, estvar.filter, landarea, and chng_type}

\item{attribute_nbrd}{EVALIDator attribute number for denominator estimation variable.}

\item{strFilter}{an additional filter to add.}

\item{EVALIDATOR_POP_ESTIMATE}{table with attribute numbers (if attribute_nbr is NULL)}
}
\description{
Get estimate from EVALIDator API
}
