% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Igetdata_arcpullr.R
\name{getdistrict}
\alias{getdistrict}
\title{get District}
\usage{
getdistrict(forest = NULL, region = NULL)
}
\arguments{
\item{forest}{Forest Name}

\item{region}{NFS region}
}
\description{
get District
}
