% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEVALIDator_compare.R
\name{getFIESTAest}
\alias{getFIESTAest}
\title{Get FIESTA estimates}
\usage{
getFIESTAest(
  evalid,
  estvar = NULL,
  estvar.filter = NULL,
  landarea = "FOREST",
  rowvar = NULL,
  colvar = NULL,
  ratio = FALSE,
  ratiotype = "PERACRE",
  estvard = NULL,
  estvard.filter = NULL,
  chng_type = "total",
  chng_measurements = "both",
  woodland = "Y",
  defaultVars = FALSE,
  dsn = NULL,
  conn = NULL
)
}
\arguments{
\item{evalid}{6 digit Evaluation ID}

\item{estvar}{estimation variable}

\item{estvar.filter}{estimation filter ('TREECLCD = 2', 'STATUSCD = 2', 'STATUSCD = 1') Similar to EVALIDator estn_type ('GS'-growing-stock; 'STGD'-standing-dead; 'AL'-all live)}

\item{landarea}{('FOREST', 'TIMBERLAND', 'ALL'). Similar to EVALIDator land_basis ('Forest land', Timberland', 'All land')}

\item{rowvar}{row domain to estimate by}

\item{colvar}{column domain to estimate by}

\item{ratio}{if ratio estimation}

\item{ratiotype}{type of ratio ('PERACRE', 'PERTREE')}

\item{estvard}{estimation variable for denominator}

\item{estvard.filter}{estimation filter for denominator ('TREECLCD = 2', 'STATUSCD = 2', 'STATUSCD = 1') Similar to EVALIDator estd_type ('GS'-growing-stock; 'STGD'-standing-dead; 'AL'-all live)}

\item{chng_type}{c('total', 'annual')}

\item{chng_measurements}{('both', 'either')}

\item{woodland}{('Y', 'N', 'only') - whether to include woodland tree species}

\item{defaultVars}{Bool. Whether or not to return default variables.}

\item{dsn}{data source name of SQLite database with EVALIDATOR_POP_ESTIMATE table (if attribute_nbr is NULL)}

\item{conn}{open connection of SQLite database with EVALIDATOR_POP_ESTIMATE table (if attribute_nbr is NULL)}
}
\description{
Get FIESTA estimates
}
