% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Igetdata_arcpullr.R
\name{getfiresEDW}
\alias{getfiresEDW}
\title{get fires EDW}
\usage{
getfiresEDW(fireyear, minacres = NULL)
}
\arguments{
\item{fireyear}{Year}

\item{minacres}{Minimum acres}
}
\description{
get fires EDW
}
