% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAnalysis-package.R
\docType{data}
\name{ref_unit}
\alias{ref_unit}
\title{Reference table - 20-year stand age classes(STDAGE).}
\format{
A dataframe with 3 columns: STATECD, VALUE, and MEANING
}
\source{
Imported from comma-delimited file.
}
\description{
Table with min (MIN), and stand age class name (STDAGECLNM).
}
\keyword{datasets}
