% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabgrp.R
\name{tabgrp}
\alias{tabgrp}
\title{tabgrp function}
\usage{
tabgrp(
  GBpopdat = NULL,
  esttype,
  unitvar = "ESTN_UNIT",
  rowvar,
  colvar = NULL,
  landarea = "ALL",
  pcfilter = NULL,
  estvar = NULL,
  estvar.filter = NULL,
  colgrp = FALSE,
  colgrpcd = NULL,
  rowgrptot = TRUE,
  colgrptot = TRUE,
  sumunits = TRUE,
  rowgrp.subtot = TRUE,
  rowgrp2.subtot = FALSE,
  title.rowgrp = NULL,
  title.colgrp = NULL,
  table_opts = NULL,
  title_opts = NULL
)
}
\arguments{
\item{GBpopdat}{GB population data}

\item{esttype}{Estimation Type}

\item{unitvar}{Unit Variable}

\item{rowvar}{Row Variable}

\item{colvar}{Col Variable}

\item{landarea}{landarea filter}

\item{pcfilter}{plot/cond filter}

\item{estvar}{Estimation Variable}

\item{estvar.filter}{Estimation Variable Filter}

\item{colgrp}{colgrp}

\item{colgrpcd}{colgrpcd}

\item{rowgrptot}{rowgrptot}

\item{colgrptot}{colgrptot}

\item{sumunits}{sumunits}

\item{rowgrp.subtot}{rowgrp.subtot}

\item{rowgrp2.subtot}{rowgrp2.subtot}

\item{title.rowgrp}{title.rowgrp}

\item{title.colgrp}{title.colgrp}

\item{table_opts}{table options}

\item{title_opts}{title options}
}
\value{
list
}
\description{
tabgrp function
}
\author{
Tracey S. Frescino
}
\keyword{data}
