% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ixlsx_functions.R, R/write2xlsx.R,
%   R/xlsx_desc.R
\name{pcheck.xlsx}
\alias{pcheck.xlsx}
\alias{wrapSE}
\alias{setCells}
\alias{write2xlsx}
\alias{xlsx_desc}
\title{xlsx Functions}
\usage{
pcheck.xlsx(
  wbnm,
  savewb = TRUE,
  outfn = NULL,
  outfolder = NULL,
  outfn.date = TRUE,
  overwrite = FALSE
)

wrapSE(x)

setCells(
  datsheet,
  estgrp,
  psegrp,
  nbrgrps = 1,
  startrow = 1,
  endrow = NULL,
  subtotal = TRUE,
  totcols,
  estcols,
  psecols,
  esttab.style,
  psetab.style,
  rnames = rnames,
  norname = FALSE,
  allin1 = FALSE,
  addSEcol = FALSE
)

write2xlsx(
  esttab,
  psetab = NULL,
  wbnm = NULL,
  sheetnm = NULL,
  tabtitle = NULL,
  subtitle = NULL,
  rowgrp = FALSE,
  rowgrp2 = FALSE,
  colgrp = FALSE,
  rowgrpnm = NULL,
  rowgrp2nm = NULL,
  rowgrp.subtot = FALSE,
  rowgrp2.subtot = FALSE,
  subtotvals = NULL,
  colgrp.subtot = FALSE,
  cnames = NULL,
  rowtotal = TRUE,
  coltotal = TRUE,
  rowtottxt = "Total",
  coltottxt = "Total",
  estnm = "Estimate",
  psenm = "Percent Sampling Error",
  title.colvar = NULL,
  title.rowvar = NULL,
  title.fontsize = 12,
  fill = TRUE,
  title.rowgrp = NULL,
  estround = 1,
  pseround = 2,
  allin1 = FALSE,
  addSEcol = FALSE,
  footnote1 = NULL,
  footnote2 = NULL,
  footnote3 = NULL,
  footnote4 = NULL,
  savewb = TRUE,
  cellWidth.grp = 6,
  cellwidth = 11,
  cellwidthSE = NULL,
  psenm.out = "PSE",
  outfn = NULL,
  outfolder = NULL,
  outfn.date = FALSE,
  overwrite = FALSE,
  autosize = FALSE
)
}
\description{
These functions use the xlsx R package to write and manipulate data.
}
\author{
Tracey S. Frescino
}
\keyword{internal}
